Extract test functions to test_base
