+++ src/operator/controllers/intents_reconcilers/networkpolicy/builders/internet_network_policy_test.go
@@ -5,18 +5,12 @@ import (
 	otterizev1alpha3 "github.com/otterize/intents-operator/src/operator/api/v1alpha3"
 	"github.com/otterize/intents-operator/src/operator/controllers/intents_reconcilers"
 	"github.com/otterize/intents-operator/src/operator/controllers/intents_reconcilers/consts"
-	mocks "github.com/otterize/intents-operator/src/operator/controllers/intents_reconcilers/mocks"
-	"github.com/otterize/intents-operator/src/operator/controllers/intents_reconcilers/networkpolicy"
-	"github.com/otterize/intents-operator/src/operator/effectivepolicy"
-	"github.com/otterize/intents-operator/src/shared/testbase"
 	"github.com/sirupsen/logrus"
-	"github.com/spf13/viper"
 	"github.com/stretchr/testify/suite"
 	"go.uber.org/mock/gomock"
 	v1 "k8s.io/api/networking/v1"
 	apierrors "k8s.io/apimachinery/pkg/api/errors"
 	metav1 "k8s.io/apimachinery/pkg/apis/meta/v1"
-	"k8s.io/apimachinery/pkg/runtime"
 	"k8s.io/apimachinery/pkg/types"
 	"k8s.io/apimachinery/pkg/util/intstr"
 	ctrl "sigs.k8s.io/controller-runtime"
@@ -26,11 +20,8 @@ import (
 )
 
 type InternetNetworkPolicyReconcilerTestSuite struct {
-	testbase.MocksSuiteBase
-	Builder               *InternetEgressRulesBuilder
-	Reconciler            *networkpolicy.Reconciler
-	externalNetpolHandler *mocks.MockexternalNetpolHandler
-	EPIntentsReconciler   *intents_reconcilers.ServiceEffectivePolicyIntentsReconciler
+	RulesBuilderTestSuiteBase
+	Builder *InternetEgressRulesBuilder
 }
 
 func init() {
@@ -38,66 +29,16 @@ func init() {
 }
 
 func (s *InternetNetworkPolicyReconcilerTestSuite) SetupTest() {
-	s.MocksSuiteBase.SetupTest()
-	s.externalNetpolHandler = mocks.NewMockexternalNetpolHandler(s.Controller)
-	restrictToNamespaces := make([]string, 0)
-
-	scheme := &runtime.Scheme{}
+	s.RulesBuilderTestSuiteBase.SetupTest()
 	s.Builder = NewInternetEgressRulesBuilder()
-	s.Reconciler = networkpolicy.NewReconciler(s.Client,
-		scheme,
-		s.externalNetpolHandler,
-		restrictToNamespaces,
-		true,
-		true,
-		nil,
-		[]networkpolicy.EgressRuleBuilder{s.Builder})
-
-	EPIntentsReconciler := effectivepolicy.NewGroupReconciler(s.Client,
-		scheme, s.Reconciler)
-	s.EPIntentsReconciler = intents_reconcilers.NewServiceEffectiveIntentsReconciler(s.Client,
-		scheme, EPIntentsReconciler)
-
 	s.Builder.Recorder = s.Recorder
-	EPIntentsReconciler.InjectableRecorder.Recorder = s.Recorder
-	s.EPIntentsReconciler.Recorder = s.Recorder
+	s.Reconciler.AddEgressRuleBuilder(s.Builder)
+
 }
 
 func (s *InternetNetworkPolicyReconcilerTestSuite) TearDownTest() {
-	viper.Reset()
+	s.RulesBuilderTestSuiteBase.TearDownTest()
 	s.Builder = nil
-	s.externalNetpolHandler = nil
-	s.MocksSuiteBase.TearDownTest()
-}
-
-func (s *InternetNetworkPolicyReconcilerTestSuite) expectGetAllEffectivePolicies(clientIntents []otterizev1alpha3.ClientIntents) {
-	var intentsList otterizev1alpha3.ClientIntentsList
-
-	s.Client.EXPECT().List(gomock.Any(), &intentsList).DoAndReturn(func(_ context.Context, intents *otterizev1alpha3.ClientIntentsList, _ ...any) error {
-		intents.Items = append(intents.Items, clientIntents...)
-		return nil
-	})
-
-	// create service to ClientIntents pointing to it
-	services := make(map[string][]otterizev1alpha3.ClientIntents)
-	for _, clientIntent := range clientIntents {
-		for _, intentCall := range clientIntent.GetCallsList() {
-			server := otterizev1alpha3.GetFormattedOtterizeIdentity(intentCall.GetTargetServerName(), intentCall.GetTargetServerNamespace(clientIntent.Namespace))
-			services[server] = append(services[server], clientIntent)
-		}
-	}
-
-	matchFieldsPtr := &client.MatchingFields{}
-	s.Client.EXPECT().List(
-		gomock.Any(),
-		&otterizev1alpha3.ClientIntentsList{},
-		gomock.AssignableToTypeOf(matchFieldsPtr),
-	).DoAndReturn(func(_ context.Context, intents *otterizev1alpha3.ClientIntentsList, args ...any) error {
-		matchFields := args[0].(*client.MatchingFields)
-		intents.Items = services[(*matchFields)[otterizev1alpha3.OtterizeFormattedTargetServerIndexField]]
-		return nil
-	}).AnyTimes()
-
 }
 
 func (s *InternetNetworkPolicyReconcilerTestSuite) TestCreateNetworkPolicySingleEndpoint() {
@@ -670,80 +611,6 @@ func (s *InternetNetworkPolicyReconcilerTestSuite) TestRemoveOrphanNetworkPolicy
 	res, err := s.EPIntentsReconciler.Reconcile(context.Background(), req)
 	s.NoError(err)
 	s.Empty(res)
-}
-
-func (s *InternetNetworkPolicyReconcilerTestSuite) ignoreRemoveOrphan() {
-	labelSelector := metav1.LabelSelector{MatchExpressions: []metav1.LabelSelectorRequirement{
-		{
-			Key:      otterizev1alpha3.OtterizeNetworkPolicy,
-			Operator: metav1.LabelSelectorOpExists,
-		},
-		{
-			Key:      otterizev1alpha3.OtterizeNetworkPolicyExternalTraffic,
-			Operator: metav1.LabelSelectorOpDoesNotExist,
-		},
-		{
-			Key:      otterizev1alpha3.OtterizeNetworkPolicyServiceDefaultDeny,
-			Operator: metav1.LabelSelectorOpDoesNotExist,
-		},
-	}}
-	selector, err := metav1.LabelSelectorAsSelector(&labelSelector)
-	s.Require().NoError(err)
-
-	s.Client.EXPECT().List(gomock.Any(), gomock.Eq(&v1.NetworkPolicyList{}), gomock.Any()).DoAndReturn(
-		func(ctx context.Context, list client.ObjectList, opts *client.ListOptions) error {
-			s.Require().NotNil(opts)
-			s.Require().NotNil(opts.LabelSelector)
-			s.Require().Equal(selector.String(), opts.LabelSelector.String())
-			return nil
-		})
-
-	s.ignoreRemoveDeprecatedPolicies()
-}
-
-func (s *InternetNetworkPolicyReconcilerTestSuite) ignoreRemoveDeprecatedPolicies() {
-	deprecatedLabels := []string{otterizev1alpha3.OtterizeEgressNetworkPolicy, otterizev1alpha3.OtterizeSvcEgressNetworkPolicy, otterizev1alpha3.OtterizeInternetNetworkPolicy, otterizev1alpha3.OtterizeSvcNetworkPolicy}
-	for _, label := range deprecatedLabels {
-		selectorRequirement := metav1.LabelSelectorRequirement{
-			Key:      label,
-			Operator: metav1.LabelSelectorOpExists,
-		}
-		selector, err := metav1.LabelSelectorAsSelector(&metav1.LabelSelector{MatchExpressions: []metav1.LabelSelectorRequirement{
-			selectorRequirement,
-		}})
-
-		s.Require().NoError(err)
-
-		s.Client.EXPECT().List(gomock.Any(), gomock.Eq(&v1.NetworkPolicyList{}), &client.ListOptions{LabelSelector: selector}).Return(nil).AnyTimes()
-	}
-}
-
-func (s *InternetNetworkPolicyReconcilerTestSuite) expectRemoveOrphanFindsPolicies(netpols []v1.NetworkPolicy) {
-	labelSelector := metav1.LabelSelector{MatchExpressions: []metav1.LabelSelectorRequirement{
-		{
-			Key:      otterizev1alpha3.OtterizeNetworkPolicy,
-			Operator: metav1.LabelSelectorOpExists,
-		},
-		{
-			Key:      otterizev1alpha3.OtterizeNetworkPolicyExternalTraffic,
-			Operator: metav1.LabelSelectorOpDoesNotExist,
-		},
-		{
-			Key:      otterizev1alpha3.OtterizeNetworkPolicyServiceDefaultDeny,
-			Operator: metav1.LabelSelectorOpDoesNotExist,
-		},
-	}}
-	selector, err := metav1.LabelSelectorAsSelector(&labelSelector)
-	s.Require().NoError(err)
-
-	s.Client.EXPECT().List(
-		gomock.Any(), gomock.Eq(&v1.NetworkPolicyList{}), &client.ListOptions{LabelSelector: selector},
-	).DoAndReturn(
-		func(_ context.Context, netpolList *v1.NetworkPolicyList, _ ...any) error {
-			netpolList.Items = append(netpolList.Items, netpols...)
-			return nil
-		},
-	)
 }
 
 func (s *InternetNetworkPolicyReconcilerTestSuite) TestNetworkPolicyCreateEnforcementDisabled() {
